requirements.txt
streamlit
google-generativeai
gspread
google-auth-oauthlib
google-api-python-client
google-auth-httplib2